+++ R/input_f.R
@@ -564,7 +564,9 @@ add_item2 <- function(.data=NULL,input){
 
 
 modify_item <- function(list_item){
-
+  lifecycle::deprecate_stop("2.0.0", "modify_item()",
+                            details = "This function is deprecated, as it's no longer needed to run the simulation,
+                            not even when backwards = TRUE. Just write a standard code (see examples).")
   input_list_arm <- parent.frame()
 
   list2env(lapply(list_item, unname), parent.frame())
@@ -611,6 +613,10 @@ modify_item <- function(list_item){
 #'   })
 
 modify_item_seq <- function(...){
+  lifecycle::deprecate_stop("2.0.0", "modify_item_seq()",
+                            details = "This function is deprecated, as it's no longer needed to run the simulation,
+                            not even when backwards = TRUE. Just write a standard code (see examples).")
+  
   input_list_arm <- parent.frame()
   input_list <- as.list(substitute(...))[-1]
   list_out <- list()
@@ -1863,7 +1869,7 @@ ast_as_list <- function(ee) {
 }
 
 
-#' Extracts items and events by looking into assignments, modify_item, modify_item_seq, modify_event and new_event
+#' Extracts items and events by looking into assignments, modify_event and new_event
 #'
 #' @param node Relevant node within the nested AST list
 #' @param conditional_flag Boolean whether the statement is contained within a conditional statement
@@ -1873,44 +1879,108 @@ ast_as_list <- function(ee) {
 #'  
 #' @examples
 #' expr <- substitute({
-#' 
-#' a <- sum(5+7)
-#' 
-#' modify_item(list(afsa=ifelse(TRUE,"asda",NULL)))
-#' 
-#' modify_item_seq(list(
 #'   
-#'   o_other_q_gold1 = if(gold == 1) { utility } else { 0 },
+#'   a <- sum(5+7)
 #'   
-#'   o_other_q_gold2 = if(gold == 2) { utility } else { 0 },
+#'   ggplot()
 #'   
-#'   o_other_q_gold3 = if(gold == 3) { utility } else { 0 },
+#'   data.frame(x=1,b=2)
 #'   
-#'   o_other_q_gold4 = if(gold == 4) { utility } else { 0 },
+#'   list(b=5)
+#'   
+#'   a <- list(s=7)
+#'   
+#'   
+#'   j <- 6
+#'   if(TRUE){modify_event(list(j=5))}
+#'   
+#'   l <- 9
+#'   
+#'   afsa=ifelse(TRUE,"asda",NULL)
+#'   
+#'   
+#'   o_exn = o_exn + 1
+#'   
+#'   a = NULL
+#'   
+#'   b = if(a){"CZ"}else{"AW"}
+#'   
+#'   rnd_prob_exn_sev = runif(1)
+#'   
+#'   exn_sev = rnd_prob_exn_sev <= p_sev
+#'   
+#'   o_exn_mod = o_exn_mod + if(exn_sev) { 0 } else { 1 }
+#'   
+#'   o_exn_sev = o_exn_sev + if(exn_sev) { 1 } else { 0 }
+#'   
+#'   o_rec_time_without_exn = (o_exn == 0) * 1
+#'   
+#'   o_rec_time_without_exn_sev = (o_exn_sev == 0) * 1
+#'   
+#'   o_c_exn = if(exn_sev) { c_sev } else { c_mod }
+#'   
+#'   o_other_c_exn_mod = if(exn_sev) { 0 } else { c_mod }
+#'   
+#'   o_other_c_exn_sev = if(exn_sev) { c_sev } else { 0 }
+#'   
+#'   o_qloss_exn = -if(exn_sev) { q_sev } else { q_mod }
+#'   
+#'   o_other_qloss_exn_mod = -if(exn_sev) { 0 } else { q_mod }
+#'   
+#'   o_other_qloss_exn_sev = -if(exn_sev) { q_sev } else { 0 }
+#'   
+#'   o_qloss_cg_exn = -if(exn_sev) { q_cg_sev } else { q_cg_mod }
+#'   
+#'   o_other_qloss_cg_exn_mod = -if(exn_sev) { 0 } else { q_cg_mod }
+#'   
+#'   o_other_qloss_cg_exn_sev = -if(exn_sev) { q_cg_sev } else { 0 }
+#'   
+#'   o_q = utility
+#'   
+#'   o_other_q_gold1 = if(gold == 1) { utility } else { 0 }
+#'   
+#'   o_other_q_gold2 = if(gold == 2) { utility } else { 0 }
+#'   
+#'   o_other_q_gold3 = if(gold == 3) { utility } else { 0 }
+#'   
+#'   o_other_q_gold4 = if(gold == 4) { utility } else { 0 }
 #'   
 #'   o_other_q_on_dup = if(on_dup) { utility } else { 0 }
-#'  
-#' ))
-#' 
-#' if(a==1){
-#'   modify_item(list(a=list(6+b)))
 #'   
-#'   modify_event(list(e_exn = curtime + 14 / days_in_year + qexp(rnd_exn, r_exn)))
-#' } else{
-#'   modify_event(list(e_exn = curtime + 14 / days_in_year + qexp(rnd_exn, r_exn)))
-#'   if(a>6){
-#'     modify_item(list(a=8))
+#'   n_exn = n_exn + 1
+#'   
+#'   n_exn_mod = n_exn_mod + (1 - exn_sev)
+#'   
+#'   n_exn_sev = n_exn_sev + exn_sev
+#'   
+#'   u_adj_exn_lt = u_adj_exn_lt + if(exn_sev) { u_adj_sev_lt } else { u_adj_mod_lt }
+#'   
+#'   utility = u_gold - u_adj_exn_lt - u_mace_lt
+#'   
+#'   o_rec_utility = utility
+#'   
+#'   rnd_exn = runif(1)
+#'   
+#'   
+#'   if(a==1){
+#'     a=list(6+b)
+#'     
+#'     modify_event(list(e_exn = curtime + 14 / days_in_year + qexp(rnd_exn, r_exn)))
+#'   } else{
+#'     modify_event(list(e_exn = curtime + 14 / days_in_year + qexp(rnd_exn, r_exn)))
+#'     if(a>6){
+#'       a=8
+#'     }
+#'     
 #'   }
 #'   
-#' }
-#' 
-#' 
-#' if (sel_resp_incl == 1 & on_dup == 1) {
 #'   
-#'   modify_event(list(e_response = curtime, z = 6))
+#'   if (sel_resp_incl == 1 & on_dup == 1) {
+#'     
+#'     modify_event(list(e_response = curtime, z = 6))
+#'     
+#'   }
 #'   
-#' }
-#' 
 #' })
 #' 
 #'