+++ NEWS.md
@@ -1,3 +1,7 @@
+# WARDEN 1.3.5
+* Accumulation backwards now uses active bindings to recognize which ongoing outputs are being modified.
+* Deprecated modify_item_seq, modify_item as they are no longer needed to run even if backwards = TRUE. Removed from all examples.
+
 # WARDEN 1.3.4
 * Unit tests for shared_inputs and run_sim_parallel added.
 * add_item and add_item2 have now been integrated into add_item, with both behaviors being accepted. This means load_inputs has been overwritten with the load_inputs2 form.